Consolidate Platform Specifics; legacy colors working for iOS Picker and DatePicker
