+++ Xamarin.Forms.Controls/GalleryPages/VisualStateManagerGalleries/ButtonDisabledStatesGallery.xaml
@@ -55,8 +55,8 @@
 				whatever colors are manually set for it -->
 				<Label Text="The Button below has the Legacy Color Behavior disabled; it will stick with whatever colors are set, regardless of state"/>
 				<Button Text="Button" x:Name="Button1" TextColor="Blue" BackgroundColor="Bisque"
-						androidSpecific:Button.IsLegacyColorModeEnabled="False"
-						iosSpecific:Button.IsLegacyColorModeEnabled="False"
+						androidSpecific:VisualElement.IsLegacyColorModeEnabled="False"
+						iosSpecific:VisualElement.IsLegacyColorModeEnabled="False"
 						/>
 				<Button Text="Toggle IsEnabled" x:Name="Button1_Toggle" Clicked="Button1_Toggle_OnClicked" />
