+++ Xamarin.Forms.Controls/GalleryPages/VisualStateManagerGalleries/DatePickerDisabledStatesGallery.xaml
@@ -54,7 +54,8 @@
 					whatever colors are manually set for it -->
 				<Label Text="The DatePicker below has the Legacy Color Behavior disabled; it will stick with whatever colors are set, regardless of state"/>
 				<DatePicker x:Name="Picker1" TextColor="Blue" BackgroundColor="Bisque" 
-							androidSpecific:DatePicker.IsLegacyColorModeEnabled="False">
+							androidSpecific:VisualElement.IsLegacyColorModeEnabled="False"
+							iosSpecific:VisualElement.IsLegacyColorModeEnabled="False">
 				</DatePicker>
 				<Button Text="Toggle IsEnabled" x:Name="Button1" Clicked="Button1_OnClicked" />
