+++ Xamarin.Forms.Controls/GalleryPages/VisualStateManagerGalleries/TimePickerDisabledStatesGallery.xaml
@@ -54,7 +54,7 @@
 					whatever colors are manually set for it -->
 				<Label Text="The TimePicker below has the Legacy Color Behavior disabled; it will stick with whatever colors are set, regardless of state"/>
 				<TimePicker x:Name="Picker1" TextColor="Blue" BackgroundColor="Bisque" 
-					androidSpecific:TimePicker.IsLegacyColorModeEnabled="False">
+					androidSpecific:VisualElement.IsLegacyColorModeEnabled="False">
 				</TimePicker>
 				<Button Text="Toggle IsEnabled" x:Name="Button1" Clicked="Button1_OnClicked" />
