+++ Xamarin.Forms.Core/PlatformConfiguration/AndroidSpecific/VisualElement.cs
@@ -1,12 +1,41 @@
 ﻿namespace Xamarin.Forms.PlatformConfiguration.AndroidSpecific
 {
-	using FormsElement = Forms.Entry;
+	using FormsElement = Forms.VisualElement;
 
-	public static class Entry
+	public static class VisualElement
     {
-	    #region IsLegacyColorModeEnabled
+		#region Elevation
 
-	    public static readonly BindableProperty IsLegacyColorModeEnabledProperty =
+		public static readonly BindableProperty ElevationProperty =
+			BindableProperty.Create("Elevation", typeof(float?),
+				typeof(FormsElement));
+
+		public static float? GetElevation(FormsElement element)
+		{
+			return (float?)element.GetValue(ElevationProperty);
+		}
+
+		public static void SetElevation(FormsElement element, float? value)
+		{
+			element.SetValue(ElevationProperty, value);
+		}
+
+		public static float? GetElevation(this IPlatformElementConfiguration<Android, FormsElement> config)
+		{
+			return GetElevation(config.Element);
+		}
+
+		public static IPlatformElementConfiguration<Android, FormsElement> SetElevation(this IPlatformElementConfiguration<Android, FormsElement> config, float? value)
+		{
+			SetElevation(config.Element, value);
+			return config;
+		}
+
+		#endregion
+
+		#region IsLegacyColorModeEnabled
+
+		public static readonly BindableProperty IsLegacyColorModeEnabledProperty =
 		    BindableProperty.CreateAttached("IsLegacyColorModeEnabled", typeof(bool),
 			    typeof(FormsElement), true);
