+++ Xamarin.Forms.Core/PlatformConfiguration/iOSSpecific/VisualElement.cs
@@ -29,5 +29,35 @@ namespace Xamarin.Forms.PlatformConfiguration.iOSSpecific
 			SetBlurEffect(config.Element, value);
 			return config;
 		}
+
+		#region IsLegacyColorModeEnabled
+
+		public static readonly BindableProperty IsLegacyColorModeEnabledProperty =
+			BindableProperty.CreateAttached("IsLegacyColorModeEnabled", typeof(bool),
+				typeof(FormsElement), true);
+
+		public static bool GetIsLegacyColorModeEnabled(BindableObject element)
+		{
+			return (bool)element.GetValue(IsLegacyColorModeEnabledProperty);
+		}
+
+		public static void SetIsLegacyColorModeEnabled(BindableObject element, bool value)
+		{
+			element.SetValue(IsLegacyColorModeEnabledProperty, value);
+		}
+
+		public static bool GetIsLegacyColorModeEnabled(this IPlatformElementConfiguration<iOS, FormsElement> config)
+		{
+			return (bool)config.Element.GetValue(IsLegacyColorModeEnabledProperty);
+		}
+
+		public static IPlatformElementConfiguration<iOS, FormsElement> SetIsLegacyColorModeEnabled(
+			this IPlatformElementConfiguration<iOS, FormsElement> config, bool value)
+		{
+			config.Element.SetValue(IsLegacyColorModeEnabledProperty, value);
+			return config;
+		}
+
+		#endregion
 	}
 }
