+++ Xamarin.Forms.Platform.iOS/Renderers/PickerRenderer.cs
@@ -12,6 +12,7 @@ namespace Xamarin.Forms.Platform.iOS
 		UIPickerView _picker;
 		UIColor _defaultTextColor;
 		bool _disposed;
+		bool _useLegacyColorManagement;
 
 		IElementController ElementController => Element as IElementController;
 
@@ -50,6 +51,8 @@ namespace Xamarin.Forms.Platform.iOS
 					entry.InputAccessoryView = toolbar;
 
 					_defaultTextColor = entry.TextColor;
+					
+					_useLegacyColorManagement = e.NewElement.UseLegacyColorManagement();
 
 					SetNativeControl(entry);
 				}
@@ -158,10 +161,13 @@ namespace Xamarin.Forms.Platform.iOS
 		{
 			var textColor = Element.TextColor;
 
-			if (textColor.IsDefault || !Element.IsEnabled)
+			if (textColor.IsDefault || (!Element.IsEnabled && _useLegacyColorManagement))
 				Control.TextColor = _defaultTextColor;
 			else
 				Control.TextColor = textColor.ToUIColor();
+
+			// HACK This forces the color to update; there's probably a more elegant way to make this happen
+			Control.Text = Control.Text;
 		}
 
 		protected override void Dispose(bool disposing)
